:ok_hand: attend to some pr commentary
- guard against null paging options
- rename shuffle function to reflect algorithm
